review comments; junit not done yet
